What a mess Dungeon enemies work most of the time now. Switches etc need to be added to subgroup consideration
