+++ EnemizerLibrary/EnemizerLibrary.csproj
@@ -84,9 +84,10 @@
     <Compile Include="DungeonsRoomsSprites.cs" />
     <Compile Include="DungeonsSpritesRandomization.cs" />
     <Compile Include="EnemyRandomizer\DungeonEnemyRandomizer.cs" />
+    <Compile Include="EnemyRandomizer\GroupSubsetPossibleSprite.cs" />
     <Compile Include="EnemyRandomizer\Room.cs" />
     <Compile Include="EnemyRandomizer\RoomConstants.cs" />
-    <Compile Include="EnemyRandomizer\Sprite.cs" />
+    <Compile Include="EnemyRandomizer\DungeonSprite.cs" />
     <Compile Include="EnemyRandomizer\SpriteConstants.cs" />
     <Compile Include="EnemyRandomizer\SpriteGroup.cs" />
     <Compile Include="Extensions.cs" />
