+++ EnemizerLibrary/EnemyRandomizer/RoomConstants.cs
@@ -716,6 +716,14 @@ namespace EnemizerLibrary
             R268_MimicCave
         };//46 on 2
 
-
+        public static readonly int[] RandomizeRooms =
+        {
+            2,4,9,10,11,14,17,19,21,22,23,25,26,27,30,31,33,34,36,38,39,40,42,43,46,49,50,52,53,54,55,56,57,58,59,
+            60,61,62,63,64,65,66,67,68,69,70,73,74,75,76,78,80,81,82,83,84,85,86,87,88,89,91,92,93,94,95,96,97,98,99,
+            100,101,102,103,104,106,107,109,110,113,114,115,116,117,118,119,123,124,125,126,127,128,129,130,131,132,133,
+            135,139,140,141,142,145,146,147,149,151,152,153,155,156,157,158,159,160,161,165,166,167,168,169,170,171,174,
+            175,176,177,178,179,182,183,184,186,187,188,190,192,193,194,195,196,197,201,203,204,206,208,209,210,213,214,
+            216,217,218,219,220,223,224,228,232,238,239,240,249,251,235,254,263,264,268,269,291,267,185,286,181,150
+        };
     }
 }
